Ideal objects

Primitive solids

ConstructiveGeometry.square — Function

square(size; origin, center=false)

An axis-parallel square or rectangle with given size (scalar or vector of length 2).

Missing docstring.

Missing docstring for ConstructiveGeometry.circle. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ConstructiveGeometry.cube. Check Documenter's build log for details.

ConstructiveGeometry.cone — Function

cone(h, r)
cone(apex, r)

Circular cone.

cone(h, shape)
cone(h)*shape
cone(apex, shape)
cone(apex)*shape

Cone with arbitrary base.

ConstructiveGeometry.cylinder — Function

cylinder(h, r1, r2 [, center=false])
cylinder(h, (r1, r2) [, center=false])
cylinder(h, r [, center=false])

Warning: cylinder(h,r) is interpreted as cylinder(h,r,r), not (h,r,0) as in OpenSCAD.

Missing docstring.

Missing docstring for ConstructiveGeometry.sphere. Check Documenter's build log for details.

ConstructiveGeometry.polygon — Function

polygon(path)

Missing docstring.

Missing docstring for ConstructiveGeometry.surface. Check Documenter's build log for details.

Transformations

All transformations accept two possible syntaxes:

    transform(parameters, solid1, solid2, ...)
    transform(parameters) * solid1

The second, multiplicative form allows easy chaining of transformations:

    transform1(param1) * transform2(param2) * solid

This form may also be applied to several solids by either wrapping them in a union, or equivalently, by applying it to a Vector of ConstructiveGeometry:

    transform(parameters) * [ solid1, solid2, ... ]

Affine transformations

ConstructiveGeometry.mult_matrix — Function

mult_matrix(a, [center=c], solid...)
mult_matrix(a, b, solid...)
mult_matrix(a, b) * solid

Represents the affine operation x -> a*x + b.

Extended help

Types of `mult_matrix` parameters

The precise type of parameters a and b is not specified. Usually, a will be a matrix and b a vector, but this is left open on purpose; for instance, a can be a scalar (for a scaling). Any types so that a * Vector + b is defined will be accepted.

Conversion to a matrix will be done when converting to OpenSCAD format.

Matrix multiplication

Chained mult_matrix operations will be combined into a single operation when possible. This saves time: multiple (3 × n) matrix multiplications are replaced by (3 × 3) multiplications, followed by a single (3 × n).

ConstructiveGeometry.translate — Function

translate(v, s...)
translate(v) * s

Translates solids s... by vector v.

ConstructiveGeometry.scale — Function

scale(a, s...; center=0)
scale(a; center=0) * s

Scales solids s by factor a. If center is given then this will be the invariant point.

a may also be a vector, in which case coordinates will be multiplied by the associated diagonal matrix.

ConstructiveGeometry.rotate — Function

rotate(θ, {center=center}, {solid...})
rotate(θ, axis=axis, {center=center}, {solid...})

Rotation around the Z-axis (in trigonometric direction, i.e. counter-clockwise).

rotate((θ,φ,ψ), {center=center}, {solid...})

Rotation given by Euler angles (ZYX; same ordering as OpenSCAD).

ConstructiveGeometry.mirror — Function

mirror(v, s...; center=0)
mirror(v; center=0) * s

Reflection with axis given by the hyperplane normal to v. If center is given, then the affine hyperplane through this point will be used.

TODO: ConstructiveGeometry.project, ConstructiveGeometry.cut.

2d drawing

ConstructiveGeometry.offset — Function

offset(r, solid...; kwargs...)
offset(r; kwargs...) * solid

Offsets by given radius.

ends=:round|:square|:butt|:loop
join=:round|:miter|:square

ConstructiveGeometry.draw — Function

draw(path, width; kwargs)
ends = :loop|:butt|:square|:round
join = :square|:round|:miter
miter_limit = 2.0

Draws a path of given width.

Extrusion

ConstructiveGeometry.linear_extrude — Function

linear_extrude(h, s...)
linear_extrude(h) * s...

Linear extrusion to height h.

ConstructiveGeometry.rotate_extrude — Function

rotate_extrude([angle = 360°], solid...)
rotate_extrude([angle = 360°]) * solid

Similar to OpenSCAD's rotate_extrude primitive.

Missing docstring.

Missing docstring for ConstructiveGeometry.path_extrude. Check Documenter's build log for details.

Inserting metadata

A couple of transformations attach metadata to objects. These are defined using the same base types as affine transforms and can therefore be applied using the same syntax, i.e. either as transform(parameters, s...) or as a product transform(parameters) * s.

ConstructiveGeometry.color — Function

color(c::Colorant, s...)
color(c::AbstractString, s...)
color(c::AbstractString, α::Real, s...)
color(c) * s...

Colors objects s... in the given color.

ConstructiveGeometry.set_parameters — Function

set_parameters(;accuracy, precision, symmetry, ε, type) * solid...

A transformation which passes down the specified parameter values to its child. Roughly similar to setting $fs and $fa in OpenSCAD.

The set_parameters transformation allows attaching arbitrary metadata. This is on purpose (although there currently exists no easy way for an user to recover these metadata while meshing an object).

Defining a custom transformation

Operations

Missing docstring.

Missing docstring for union(::AbstractGeometry,::AbstractGeometry). Check Documenter's build log for details.

Missing docstring.

Missing docstring for intersect(::AbstractGeometry,::AbstractGeometry). Check Documenter's build log for details.

Missing docstring.

Missing docstring for setdiff(::AbstractGeometry,::AbstractGeometry). Check Documenter's build log for details.

ConstructiveGeometry.hull — Function

hull(s::AbstractGeometry...)
	hull(s::AbstractGeometry{2} | StaticVector{2}...)

Represents the convex hull of given solids (and, possibly, points).

ConstructiveGeometry.minkowski — Function

minkowski(s::AbstractGeometry...)

Represents the Minkowski sum of given solids.