Mesh I/O

STL output

ConstructiveGeometry.stl — Function

stl(file, object; options...)

Outputs an STL description of object to the given file (string or IO). Optional kwargs are the same as for the mesh function or the set_parameters object.

TODO: other file formats (.ply) ?

SVG output

ConstructiveGeometry.svg — Function

svg(file, shape)

Exports 2d shape as an SVG file.

STL input

TODO.

OpenSCAD output

TODO.

OpenSCAD to Julia conversion.

Might be possible for a limited subset of OpenSCAD language. TODO.