ConstructiveGeometry.jl Documentation

Warning

This package is very much a work-in-progress. Right now only very basic functionality is available (describing geometries, and some cases for meshing). Neither interface nor code are stable for now. Any contributions are welcome!

This package provides tools for describing 3d objects in Julia. This includes both geometry functions and a syntax describing constructive geometry. This syntax is inspired by OpenSCAD, but is actual Julia code:

using ConstructiveGeometry

square(20)

linear_extrude(30) * [
  intersection(
    translate([10,0]) * circle(3),
    translate([13,0]) * circle(3),
  ),
  color("pink") * scale(2) * square(1),
]

The package is made of two parts:

a system for describing geometric objects;

back-ends for converting these objects to useful formats.

As of 2021-01, the only useable output format is an OpenSCAD file.

Other planned output formats include:

(TODO) represented graphically using one of the Julia plotting packages;

(PARTLY DONE) converted to a mesh;

(TODO) directly exported as a 2d (.svg)

or 3d file (.stl or .ply).

Quick-start

Basic example

using ConstructiveGeometry

s1 = union(
  color("pink")*
  translate([3,0])*
  scale([2,1])*
  circle(3),

  color("cyan")*
  translate([0,5])*
  square([2,3])
)

mesh(s1)

I/O

ConstructiveGeometry.include — Function

ConstructiveGeometry.include(file::AbstractString, f::Function)

Reads given file and returns the union of all top-level Geometry objects (except the results of assignments) found in the file.

#### Example: contents of file `example.jl`
C=ConstructiveGeometry.cube(1)
S=ConstructiveGeometry.square(1)
ConstructiveGeometry.circle(3)
S

julia> ConstructiveGeometry.include("example.jl")
union() {
 circle(radius=3.0);
 square(size=[1.0, 1.0], center=false);
}

ConstructiveGeometry.scad — Function

scad(filename::AbstractString, s::Geometry...)
scad(io::IO, s::Geometry)

Prints an OpenSCAD-like representation of the given solid(s).

The various scad_* functions

`scad_name(s)`

Returns, in printable form (e.g. Symbol or String), the OpenSCAD name of this object.

`scad_parameters(s)`

Returns a NamedTuple representing parameters of this object.

`scad_transform(s)`

Possible transformation prepended to the object.

`to_scad(x)`

Represents value x (number, array, etc.) in OpenSCAD format.