ConstructiveGeometry.jl Documentation

Warning

This package is very much a work-in-progress. Right now only very basic functionality is available (describing geometries, and some cases for meshing). Neither interface nor code are stable for now. Any contributions are welcome!

This package provides tools for describing 3d objects in Julia. This includes both geometry functions and a syntax describing constructive geometry. This syntax is inspired by OpenSCAD, but is actual Julia code:

using ConstructiveGeometry

Square(20)

linear_extrude(30) * [
  intersection(
    translate([10,0]) * Circle(3),
    translate([13,0]) * Circle(3),
  ),
  color("pink") * scale(2) * Square(1),
]

The package is made of two parts:

a system for describing geometric objects;

back-ends for converting these objects to useful formats.

As of 2021-01, the only useable output format is an OpenSCAD file.

Other planned output formats include:

(TODO) represented graphically using one of the Julia plotting packages;

(IN PROGRESS) converted to a mesh;

(TODO) directly exported as a 2d (.svg)

or 3d file (.stl or .ply).

Quick-start

Basic example

using ConstructiveGeometry
import ConstructiveGeometry: Square, Circle, Surface, mult_matrix, translate, scale, color

s1 = union(
  color("pink")*
  translate([3,0])*
  scale([2,1])*
  Circle(3),

  color("cyan")*
  translate([0,5])*
  Square([2,3])
)

Surface(s1)

I/O

Missing docstring.

Missing docstring for ConstructiveGeometry.include. Check Documenter's build log for details.

ConstructiveGeometry.scad — Function

scad(filename::AbstractString, s::Geometry...)
scad(io::IO, s::Geometry)

Prints an OpenSCAD-like representation of the given solid(s).

The various scad_* functions

`scad_name(s)`

Returns, in printable form (e.g. Symbol or String), the OpenSCAD name of this object.

`scad_parameters(s)`

Returns a NamedTuple representing parameters of this object.

`scad_transform(s)`

Possible transformation prepended to the object.

`to_scad(x)`

Represents value x (number, array, etc.) in OpenSCAD format.