Ideal objects

Primitive solids

ConstructiveGeometry.Square — Type

Square(size; origin, center=false)

An axis-parallel square or rectangle with given size (scalar or vector of length 2).

GeometryBasics.Circle — Type

Circle{T}

An alias for a HyperSphere of dimension 2. (i.e. HyperSphere{2, T})

Circle(radius)

A circle. Discretization is done via the accuracy and precision parameters.

ConstructiveGeometry.Cube — Type

Cube(size; origin, center=false)

A cube or parallelepiped with given size (scalar or vector of length 3).

Missing docstring.

Missing docstring for ConstructiveGeometry.Cylinder. Check Documenter's build log for details.

GeometryBasics.Sphere — Type

Sphere{T}

An alias for a HyperSphere of dimension 3. (i.e. HyperSphere{3, T})

Sphere(radius)

A sphere. Discretization is done via the accuracy and precision parameters.

ConstructiveGeometry.Polygon — Type

Polygon{T}
Polygon([point1, point2, ...])
Polygon(point1, point2, ...)

A simple, closed polygon enclosed by the given vertices.

Missing docstring.

Missing docstring for ConstructiveGeometry.Surface. Check Documenter's build log for details.

ConstructiveGeometry.NeutralSolid — Type

NeutralSolid{D,T}

A convenience type representing either an empty or full solid. This exists mostly to provide a neutral element for union() and intersect() operators, hence the name. In those cases, it is impossible to know in advance the dimension of the returned solid; hence, as an exception to the general rule, the D type parameter is either the symbol :empty or :full.

Since neutral objects are removed at compile-time from corresponding CSG operations, this should have no influence on the dimension of a top-level object.

The T type parameter is always Bool.

Transformations

All transformations accept two possible syntaxes:

    transform(parameters, solid1, solid2, ...)
    transform(parameters) * solid1

The second, multiplicative form allows easy chaining of transformations:

    transform1(param1) * transform2(param2) * solid

This form may also be applied to several solids by either wrapping them in a union, or equivalently, by applying it to a Vector of ConstructiveGeometry:

    transform(parameters) * [ solid1, solid2, ... ]

Affine transformations

ConstructiveGeometry.mult_matrix — Function

mult_matrix(a, [center=c], solid...)
mult_matrix(a, b, solid...)
mult_matrix(a, b) * solid

Represents the affine operation x -> a*x + b.

Extended help

Types of `mult_matrix` parameters

The precise type of parameters a and b is not specified. Usually, a will be a matrix and b a vector, but this is left open on purpose; for instance, a can be a scalar (for a scaling) and b can be Val(false) for a linear operation. Any types so that a * Vector + b is defined will be accepted.

Conversion to a matrix will be done when converting to OpenSCAD format.

Matrix multiplication

Chained mult_matrix operations will be combined into a single operation when possible. This saves time: multiple (3 × n) matrix multiplications are replaced by (3 × 3) multiplications, followed by a single (3 × n).

ConstructiveGeometry.translate — Function

translate(v, s...)
translate(v) * s

Translates solids s... by vector v.

ConstructiveGeometry.scale — Function

scale(a, s...; center=0)
scale(a; center=0) * s

Scales solids s by factor a. If center is given then this will be the invariant point.

a may also be a vector, in which case coordinates will be multiplied by the associated diagonal matrix.

ConstructiveGeometry.rotate — Function

rotate(θ, {center=center}, {solid...})
rotate(θ, axis=axis, {center=center}, {solid...})

Rotation around the Z-axis (in trigonometric direction, i.e. counter-clockwise).

rotate((θ,φ,ψ), {center=center}, {solid...})

Rotation given by Euler angles (ZYX; same ordering as OpenSCAD).

ConstructiveGeometry.mirror — Function

mirror(v, s...; center=0)
mirror(v; center=0) * s

Reflection with axis given by the hyperplane normal to v. If center is given, then the affine hyperplane through this point will be used.

TODO: ConstructiveGeometry.project, ConstructiveGeometry.cut.

Extrusion

ConstructiveGeometry.linear_extrude — Function

linear_extrude(h, s...)
linear_extrude(h) * s...

Linear extrusion to height h.

Missing docstring.

Missing docstring for ConstructiveGeometry.rotate_extrude. Check Documenter's build log for details.

ConstructiveGeometry.path_extrude — Function

path_extrude(path, poly, options...)

Extrudes the given polygon (a path of points forming a simple loop) along the given path. Both arguments are provided as a Vector{SVector{2}}.

Returns a Surface (defined by points and a triangulation).

Inserting metadata

A couple of transformations attach metadata to objects. These are defined using the same base types as affine transforms and can therefore be applied using the same syntax, i.e. either as transform(parameters, s...) or as a product transform(parameters) * s.

ConstructiveGeometry.color — Function

color(c::Colorant, s...)
color(c::AbstractString, s...)
color(c::AbstractString, α::Real, s...)
color(c) * s...

Colors objects s... in the given color.

ConstructiveGeometry.set_parameters — Function

set_parameters(;accuracy=accuracy, precision=precision)

A transformation which passes down the specified parameter values to its child. Roughly similar to setting $fs and $fa in OpenSCAD.

Defining a custom transformation

Operations

Base.union — Function

union(s, itrs...)
∪(s, itrs...)

Construct the union of sets. Maintain order with arrays.

Examples

julia> union([1, 2], [3, 4])
4-element Array{Int64,1}:
 1
 2
 3
 4

julia> union([1, 2], [2, 4])
3-element Array{Int64,1}:
 1
 2
 4

julia> union([4, 2], 1:2)
3-element Array{Int64,1}:
 4
 2
 1

julia> union(Set([1, 2]), 2:3)
Set{Int64} with 3 elements:
  2
  3
  1

Base.intersect — Function

intersect(s, itrs...)
∩(s, itrs...)

Construct the intersection of sets. Maintain order with arrays.

Examples

julia> intersect([1, 2, 3], [3, 4, 5])
1-element Array{Int64,1}:
 3

julia> intersect([1, 4, 4, 5, 6], [4, 6, 6, 7, 8])
2-element Array{Int64,1}:
 4
 6

julia> intersect(Set([1, 2]), BitSet([2, 3]))
Set{Int64} with 1 element:
  2

intersect(ss, others...)

Each argument is a SortedSet with the same key and order type. The return variable is a new SortedSet that is the intersection of all the sets that are input. Time: O(cn log n), where n is the total number of items in all the arguments.

intersect(g, h)

Return a graph with edges that are only in both graph g and graph h.

Implementation Notes

This function may produce a graph with 0-degree vertices. Preserves the eltype of the input graph.

Examples

julia> g1 = SimpleDiGraph([0 1 0 0 0; 0 0 1 0 0; 1 0 0 1 0; 0 0 0 0 1; 0 0 0 1 0]);

julia> g2 = SimpleDiGraph([0 1 0; 0 0 1; 1 0 0]);

julia> foreach(println, edges(intersect(g1, g2)))
Edge 1 => 2
Edge 2 => 3
Edge 3 => 1

intersect(P1::HRep, P2::HRep)

Takes the intersection of P1 and P2 $\{\, x : x \in P_1, x \in P_2 \,\}$. It is very efficient between two H-representations or between two polyhedron for which the H-representation has already been computed. However, if P1 (resp. P2) is a polyhedron for which the H-representation has not been computed yet, it will trigger a representation conversion which is costly. See the Polyhedral Computation FAQ for a discussion on this operation.

The type of the result will be chosen closer to the type of P1. For instance, if P1 is a polyhedron (resp. H-representation) and P2 is a H-representation (resp. polyhedron), intersect(P1, P2) will be a polyhedron (resp. H-representation). If P1 and P2 are both polyhedra (resp. H-representation), the resulting polyhedron type (resp. H-representation type) will be computed according to the type of P1. The coefficient type however, will be promoted as required taking both the coefficient type of P1 and P2 into account.

intersect(v::VRepresentation{T}, h::HRepElement)

Compute the intersection of v with an halfspace or hyperplane h. The method used by default is to keep the V-representation element of v that are in h and add new ones generated as the intersection between the hyperplane defining h and the segment between two adjacent V-representation elements of v that are in either sides of the hyperplane. See Lemma 3 of [FP96] for more detail on the method.

[FP96] Fukuda, K. and Prodon, A. Double description method revisited Combinatorics and computer science, Springer, 1996, 91-111

intersect(s::AbstractGeometry...)
s1 ∩ s2

Represents the intersection of given solids.

ConstructiveGeometry.difference — Function

ConstructiveGeometry.difference(s1, s2)

Represents the difference s1 ∖ s2.

ConstructiveGeometry.hull — Function

hull(s::AbstractGeometry...)

Represents the convex hull of given solids.

ConstructiveGeometry.minkowski — Function

minkowski(s::AbstractGeometry...)

Represents the Minkowski sum of given solids.