Ideal objects

Primitive solids

ConstructiveGeometry.Square — Type

Square(size; origin, center=false)

An axis-parallel square or rectangle with given size (scalar or vector of length 2).

ConstructiveGeometry.Circle — Type

Circle(radius)

A circle. Discretization is done via the accuracy and precision parameters.

ConstructiveGeometry.Cube — Type

Cube(size; origin, center=false)

A cube or parallelepiped with given size (scalar or vector of length 3).

ConstructiveGeometry.Cylinder — Type

Cylinder(h, r1, r2 [, center=false])
Cylinder(h, (r1, r2) [, center=false])
Cylinder(h, r [, center=false])

Warning: Cylinder(h,r) is interpreted as Cylinder(h,r,r), not (h,r,0) as in OpenSCAD.

ConstructiveGeometry.Sphere — Type

Sphere(radius)

A sphere. Discretization is done via the accuracy and precision parameters.

ConstructiveGeometry.Polygon — Type

Polygon{T}
Polygon([point1, point2, ...])
Polygon(point1, point2, ...)

A simple, closed polygon enclosed by the given vertices.

Missing docstring.

Missing docstring for ConstructiveGeometry.Surface. Check Documenter's build log for details.

Transformations

All transformations accept two possible syntaxes:

    transform(parameters, solid1, solid2, ...)
    transform(parameters) * solid1

The second, multiplicative form allows easy chaining of transformations:

    transform1(param1) * transform2(param2) * solid

This form may also be applied to several solids by either wrapping them in a union, or equivalently, by applying it to a Vector of ConstructiveGeometry:

    transform(parameters) * [ solid1, solid2, ... ]

Affine transformations

ConstructiveGeometry.mult_matrix — Function

mult_matrix(a, [center=c], solid...)
mult_matrix(a, b, solid...)
mult_matrix(a, b) * solid

Represents the affine operation x -> a*x + b.

Extended help

Types of `mult_matrix` parameters

The precise type of parameters a and b is not specified. Usually, a will be a matrix and b a vector, but this is left open on purpose; for instance, a can be a scalar (for a scaling) and b can be Val(false) for a linear operation. Any types so that a * Vector + b is defined will be accepted.

Conversion to a matrix will be done when converting to OpenSCAD format.

Matrix multiplication

Chained mult_matrix operations will be combined into a single operation when possible. This saves time: multiple (3 × n) matrix multiplications are replaced by (3 × 3) multiplications, followed by a single (3 × n).

ConstructiveGeometry.translate — Function

translate(v, s...)
translate(v) * s

Translates solids s... by vector v.

ConstructiveGeometry.scale — Function

scale(a, s...; center=0)
scale(a; center=0) * s

Scales solids s by factor a. If center is given then this will be the invariant point.

a may also be a vector, in which case coordinates will be multiplied by the associated diagonal matrix.

ConstructiveGeometry.rotate — Function

rotate(θ, {center=center}, {solid...})
rotate(θ, axis=axis, {center=center}, {solid...})

Rotation around the Z-axis (in trigonometric direction, i.e. counter-clockwise).

rotate((θ,φ,ψ), {center=center}, {solid...})

Rotation given by Euler angles (ZYX; same ordering as OpenSCAD).

ConstructiveGeometry.mirror — Function

mirror(v, s...; center=0)
mirror(v; center=0) * s

Reflection with axis given by the hyperplane normal to v. If center is given, then the affine hyperplane through this point will be used.

TODO: ConstructiveGeometry.project, ConstructiveGeometry.cut.

Extrusion

ConstructiveGeometry.linear_extrude — Function

linear_extrude(h, s...)
linear_extrude(h) * s...

Linear extrusion to height h.

Missing docstring.

Missing docstring for ConstructiveGeometry.rotate_extrude. Check Documenter's build log for details.

ConstructiveGeometry.path_extrude — Function

path_extrude(path, poly, options...)

Extrudes the given polygon (a path of points forming a simple loop) along the given path. Both arguments are provided as a Vector{SVector{2}}.

Returns a Surface (defined by points and a triangulation).

Inserting metadata

A couple of transformations attach metadata to objects. These are defined using the same base types as affine transforms and can therefore be applied using the same syntax, i.e. either as transform(parameters, s...) or as a product transform(parameters) * s.

ConstructiveGeometry.color — Function

color(c::Colorant, s...)
color(c::AbstractString, s...)
color(c::AbstractString, α::Real, s...)
color(c) * s...

Colors objects s... in the given color.

ConstructiveGeometry.set_parameters — Function

set_parameters(;accuracy=accuracy, precision=precision)

A transformation which passes down the specified parameter values to its child. Roughly similar to setting $fs and $fa in OpenSCAD.

Defining a custom transformation

Operations

Base.union — Method

union(s::Geometry...)
s1 ∪ s2

Represents the union of given solids.

Base.intersect — Method

intersect(s::Geometry...)
s1 ∩ s2

Represents the intersection of given solids.

ConstructiveGeometry.difference — Function

ConstructiveGeometry.difference(s1, s2)

Represents the difference s1 ∖ s2.

ConstructiveGeometry.hull — Function

hull(s::Geometry...)

Represents the convex hull of given solids.

ConstructiveGeometry.minkowski — Function

minkowski(s::Geometry...)

Represents the Minkowski sum of given solids.