Mesh I/O

Loading and saving of files can be done through the FileIO functions load and save.

File loading

Files can be loaded in this way:

using FileIO
bust = load("beethoven.ply")

The following formats should be supported: .stl (ASCII and binary) .ply (ASCII and little-endian/big-endian binary).

TODO: .dxf

File saving

Files can be saved in this way:

using FileIO
save("foo.stl", object)

The following formats are supported: .stl (ASCII) and .ply (ASCII) for volumes; .svg for shapes.

Image file formats are also supported through CairoMakie's plot functions.

OpenSCAD output

ConstructiveGeometry.scad — Function

scad(io::IO, name::Symbol, parameters, children)
scad(filename::AbstractString, s::AbstractGeometry...)
scad(io::IO, s::AbstractGeometry)

Returns, in printable form (e.g. Symbol or String), the OpenSCAD name of this object.

Deprecated

Since the constructions from this package have started to diverge from those of OpenSCAD, (and also since the possibilities for export and visualization have improved), OpenSCAD output is being deprecated and may be removed in a future version.

OpenSCAD to Julia conversion.

Might be possible for a limited subset of OpenSCAD language. TODO.